requirements.txt
aiofiles==23.2.1
altair==5.4.1
anyio==4.4.0
attrs==24.2.0
Authlib==1.2.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
exceptiongroup==1.2.2
fastapi==0.114.1
ffmpy==0.4.0
filelock==3.16.0
fonttools==4.53.1
fsspec==2024.9.0
gradio==3.50.2
gradio_client==0.6.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.24.7
idna==3.8
importlib_resources==6.4.5
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
loguru==0.7.2
MarkupSafe==2.1.5
matplotlib==3.9.2
narwhals==1.7.0
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pillow==10.4.0
planqk-service-sdk==1.4.1
pycparser==2.22
pydantic==1.9.2
pydub==0.25.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rfc3986==1.5.0
rpds-py==0.20.0
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.5
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
websockets==11.0.3
